+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/webview_cupertino.dart
@@ -45,6 +45,11 @@ class CupertinoWebView implements WebViewPlatform {
   }
 
   @override
-  Future<bool> clearCookies() =>
-      WebViewCookieManagerPlatform.instance!.clearCookies();
+  Future<bool> clearCookies() {
+    if (WebViewCookieManagerPlatform.instance == null) {
+      throw Exception(
+          'Could not clear cookies as no implementation for the WebViewCookieManagerPlatform has been registered.');
+    }
+    return WebViewCookieManagerPlatform.instance!.clearCookies();
+  }
 }